rename and consolidate new request forms into one namespace
