--- atst/forms/review.py
@@ -1,7 +1,0 @@
-from wtforms.fields import BooleanField
-
-from .forms import ValidatedForm
-
-
-class ReviewForm(ValidatedForm):
-    reviewed = BooleanField("I have reviewed this data and it is correct.")
